backend/requirements.txt
# ================================================
# OFTA Backend Dependencies
# ================================================

# FastAPI & Server
fastapi==0.115.8
uvicorn==0.34.0
python-dotenv==1.0.1
pydantic==2.9.1
pydantic_core==2.23.3
pydantic[email]

# Database
asyncpg==0.30.0
psycopg2-binary==2.9.10
SQLAlchemy==2.0.30
pandas==2.2.3

# Auth & Security
firebase-admin==6.5.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
pyjwt==2.9.0
cryptography==43.0.3

# HTTP & Requests
requests==2.32.3
httpx==0.27.2

# Utilities
pytz==2024.1
python-multipart==0.0.17

# Google Cloud
google-cloud-storage==2.14.0

# Testing (optional)
pytest==8.3.4
pytest-asyncio==0.24.0
